config minio, swagger

+++ trello/src/main/java/com/example/trello/Util/Util.java
@@ -4,7 +4,7 @@ public class Util {
     public static String randomToken() {
         String random = "";
         for (int i = 0; i < 6; i++) {
-            random = random + Math.floor(Math.random() * 10);
+            random = random + (int)(Math.floor(Math.random()*10));
         }
         return random;
     }
